authors-map to component
